pyproject.toml
[project]
name = "llmbo-bedrock"
version = "0.2.0"
description = "Large Language Model Batch Operations"
readme = "docs/README.md"
authors = [
    { name = "David Gillespie", email = "david.gillespie@digital.cabinet-office.gov.uk" }
]
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.3,<2.0",
    "pydantic>=2.10",
    "python-dotenv>=1.0.1",
]

[project.urls]
Homepage = "https://co-cddo.github.io/gds-idea-llmbo/"
Documentation = "https://co-cddo.github.io/gds-idea-llmbo/api/"
Repository = "https://github.com/co-cddo/gds-idea-llmbo.git"
Issues = "https://github.com/co-cddo/gds-idea-llmbo/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llmbo"]

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "twine>=6.1.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.3",
    "mkdocstrings[python]>=0.28.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=llmbo"
